requirements.txt
websocket-client==1.6.0
requests==2.28.2